[Feat] Add Board type along with methods and docs
This commit adds a Board type for manipulating game boards as a whole,
exposing a number of methods, requiring a few new error types, and
adding a doc.go file to explain the ideas and motivations behind the
decisions that went into the design of the Board.

Along with the game engine and the decision engine, the Board leads
the trinity of components that together define the Fearsome project.