[Refactor] Improve types for size and semantics
This commit replaces underlying ints with uint8s for a few types used
by Board, both to reduce the size of Board instances by a massive 80%,
and to more clearly express the semantics of these types.

Additionally, the convention of always using keys in struct literals
for non-anonymous structs is fully embraced, with every known violation
of this convention fixed. As it happens only one violation was found.